requirements.txt
appdirs==1.4.4
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.12.0
bs4==0.0.1
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleantext==1.1.4
click==8.1.3
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
fake-useragent==1.1.3
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parse==1.19.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
pymetamap==0.1
pyOpenSSL==23.1.1
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
skr-web-api @ file:///home/parallels/Desktop/mediscrape/server/skr_web_python_api/dist/skr_web_api-0.1-py3-none-any.whl
soupsieve==2.4
tldextract==3.4.0
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
